+++ src/test/java/study/carracing/domain/CarTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import study.carracing.util.RandomGeneratorImpl;
 
 import static org.assertj.core.api.Assertions.assertThat;
