+++ src/test/java/study/carracing/domain/RacingTest.java
@@ -1,10 +1,8 @@
-package study.carracing;
+package study.carracing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import study.carracing.domain.Racing;
-import study.carracing.domain.RacingCars;
 
 import static org.assertj.core.api.Assertions.assertThat;
